+++ ShopifySharp/WebhookPayloads/CartUpdated/CartUpdatedWebhookPriceSet.cs
@@ -1,8 +1,8 @@
 using Newtonsoft.Json;
 
-namespace ShopifySharp.Entities.CartUpdatedWebhook
+namespace ShopifySharp.WebhookPayloads.CartUpdated
 {
-    public class CartUpdatedWebhookDiscountedPriceSet
+    public class CartUpdatedWebhookPriceSet
     {
         [JsonProperty("shop_money")]
         public CartUpdatedWebhookShopMoney ShopMoney { get; set; }
