+++ ShopifySharp/WebhookPayloads/CartUpdated/CartUpdatedWebhookShopMoney.cs
@@ -1,11 +1,11 @@
 using Newtonsoft.Json;
 
-namespace ShopifySharp.Entities.CartUpdatedWebhook
+namespace ShopifySharp.WebhookPayloads.CartUpdated
 {
-    public class CartUpdatedWebhookPresentmentMoney
+    public class CartUpdatedWebhookShopMoney
     {
         [JsonProperty("amount")]
-        public string Amount { get; set; }
+        public decimal Amount { get; set; }
 
         [JsonProperty("currency_code")]
         public string CurrencyCode { get; set; }
